refactor(server): Add server documentation - PST-188
