+++ server/src/keyword/keyword.module.ts
@@ -3,6 +3,9 @@ import { KeywordController } from './keyword.controller';
 import { KeywordService } from './keyword.service';
 import { HttpModule } from '@nestjs/axios';
 
+/**
+ * This module is responsible for the '/keyword' endpoint
+ */
 @Module({
     imports: [HttpModule],
     controllers: [KeywordController],
