+++ server/src/keyword/models/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Interface for the structure of the SuggestionAPI's response
+ */
+export type SuggestionAPIResponse = [string, number][];
