--- server/src/models/index.ts
@@ -1,1 +1,0 @@
-export type SuggestionAPIResponse = [string, number][];
