+++ server/src/patents/patents.module.ts
@@ -4,6 +4,9 @@ import { HttpModule } from '@nestjs/axios';
 import { ConfigModule } from '@nestjs/config';
 import { PatentsService } from './patents.service';
 
+/**
+ * This module is responsible for the '/patents' endpoint
+ */
 @Module({
     imports: [HttpModule, ConfigModule.forRoot()],
     controllers: [PatentsController],
