+++ server/src/patents/patents.service.helper.ts
@@ -11,21 +11,28 @@ import {
 
 import * as _ from 'lodash';
 
+/**
+ * This helper class provides processing functions which can be used to convert ugly OPS responses into a
+ * simplified and good looking JSON
+ */
 export class PatentsServiceHelper {
     /**
      * Processes an OPS query and returns a much cleaner structure with the needed data
-     * @param data
-     * @param languages
+     *
+     * @param data  The data response from OPS
+     * @param languages The optional languages filter
      */
     public static processQuery(data: PatentQueryResponse, languages?: string): QueryResult {
         const searchResult = data['ops:world-patent-data']['ops:biblio-search'];
         const totalCount = data['ops:world-patent-data']['ops:biblio-search']['@total-result-count'];
 
+        // query data can be either an object or an array for some reason. Therefore if it is an object convert it to an array
         let queryData = searchResult['ops:search-result']['exchange-documents'];
         if (!(queryData instanceof Array)) {
             queryData = [queryData];
         }
 
+        // process the documents array
         let processed: OpsExchangeDocument[] | Patent[] = queryData
             .map((item) => ({
                 ...item['exchange-document'],
@@ -46,8 +53,10 @@ export class PatentsServiceHelper {
             });
         }
 
+        // process each document
         processed = processed.map((item) => PatentsServiceHelper.processPatent(item, languages));
 
+        // return patents and total number of available patents
         return {
             patents: processed,
             total: Number(totalCount),
@@ -55,9 +64,10 @@ export class PatentsServiceHelper {
     }
 
     /**
-     * Processes an OPS patent and returns the sanitized patent form
-     * @param item
-     * @param languages
+     * Processes a single OPS patent and returns the sanitized patent as a simple object
+     *
+     * @param item  The ops document
+     * @param languages The (optional) languages which should be preferred
      */
     public static processPatent(item: OpsExchangeDocument, languages = ''): Patent {
         return {
@@ -113,13 +123,14 @@ export class PatentsServiceHelper {
 
     /**
      * Attempts to get the title of a patent from the provided data
-     * @param doc
-     * @param languages
+     * @param doc   The exchange document
+     * @param languages The preferred languages
      * @private
      */
     private static getTitle(doc: OpsExchangeDocument, languages: string): string {
         let titles = doc['bibliographic-data']['invention-title'];
 
+        // some patents don't have a title for some reason
         if (!titles) {
             return 'Unknown Title';
         }
@@ -130,6 +141,7 @@ export class PatentsServiceHelper {
         }
         // filtering for the provided languages
         const titleSupported = titles.find((title) => (languages || '').includes(title['@lang']));
+
         //if the Title is not existing by default in english or one of the filtered languages return this message
         if (!titleSupported) {
             return titles[0].$;
@@ -140,16 +152,18 @@ export class PatentsServiceHelper {
 
     /**
      * Gets citations from a provided OpsExchangeDocument
-     * @param doc
+     * @param doc   The ops exchange document
      * @private
      */
     private static getCitations(doc: OpsExchangeDocument): Patent[] {
         let citations = doc['bibliographic-data']['references-cited']?.citation;
 
+        // no citations? Return an empty array!
         if (!citations || citations.length === 0) {
             return [] as Patent[];
         }
 
+        // if only one citation is present it will be an object for some weird reason
         if (!(citations instanceof Array)) {
             citations = [citations];
         }
@@ -173,8 +187,8 @@ export class PatentsServiceHelper {
     /**
      * Gets the abstract from a provided OpsExchangeDocument
      *
-     * @param doc
-     * @param languages
+     * @param doc   The exchange document
+     * @param languages The preferred languages
      * @private
      */
     private static getAbstract(doc: OpsExchangeDocument, languages: string): string {
@@ -192,6 +206,7 @@ export class PatentsServiceHelper {
 
         // filtering for the provided languages
         const abstractSupported = abstracts.find((abstract) => (languages || '').includes(abstract['@lang']));
+
         //if the abstract is not existing by default in english or one of the filtered languages return this message
         if (!abstractSupported) {
             return abstracts[0].p.$;
@@ -224,14 +239,15 @@ export class PatentsServiceHelper {
 
     /**
      * Compiles an OPS query string based on the provided data
+     *
      * @param searchTerms   The search terms which need to be added to the URL
      * @param endpoint      The target endpoint
      * @param page          The page which should be retrieved
-     * @param country          The page which should be retrieved
-     * @param date          The page which should be retrieved
-     * @param inventor
-     * @param applicant
-     * @param pageSize
+     * @param country       The country-filter
+     * @param date          The date-filter
+     * @param inventor      The inventor
+     * @param applicant     The applicant
+     * @param pageSize      The page size (default: 100)
      */
     public static getQueryString(
         searchTerms: string[],
@@ -245,6 +261,7 @@ export class PatentsServiceHelper {
     ): string {
         let queryString = `${process.env.PATENT_API_URL}`.concat(endpoint);
 
+        // filters array we will use later to add them to the query string
         const filters: string[] = [];
 
         if (searchTerms) {
@@ -267,7 +284,7 @@ export class PatentsServiceHelper {
             filters.push(`pd within "${date}"`);
         }
 
-        console.log(filters);
+        // add filters array to the query-string
         filters.forEach((filter, index) => {
             if (index === 0) {
                 queryString = queryString.concat('?q=').concat(filter);
@@ -282,7 +299,8 @@ export class PatentsServiceHelper {
 
     /**
      * Restructures an image query to the desired format
-     * @param data
+     *
+     * @param data The OPS-image query response
      */
     public static async processImageQuery(data: OpsImageQueryResponse): Promise<DocumentInformation[]> {
         return PatentsServiceHelper.getNestedArray<OpsDocumentInstance[]>(
