+++ server/src/patents/patents.service.ts
@@ -3,6 +3,7 @@ import {
     DocumentInformation,
     OpsImageQueryResponse,
     OpsPatent,
+    Patent,
     PatentFamilyResponse,
     PatentQueryResponse,
     QueryResult,
@@ -14,25 +15,33 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 
 import { PatentsServiceHelper } from './patents.service.helper';
 
+/**
+ * Contains the business logic for patent retrieval
+ */
 @Injectable()
 export class PatentsService {
     private authData: AuthResponse;
     private data: PatentQueryResponse;
 
+    // endpoints
     private static patentEndpoint = '/rest-services/published-data/search/biblio';
-    private static imageEndpoint = '/rest-services/published-data/publication/epodoc/';
+    private static epoDocEndpoint = '/rest-services/published-data/publication/epodoc/';
     private static familyEndpoint = '/rest-services/family/publication/docdb/';
 
     constructor(private readonly httpService: HttpService) {}
 
     /**
      * Queries the patent API for the patent belonging to the patentId
-     * @param patentId
+     *
+     * @param patentId  The id of the patent
      */
-    async get(patentId: string): Promise<any> {
-        const url = process.env.PATENT_API_URL.concat(PatentsService.imageEndpoint).concat(patentId).concat('/biblio');
+    async get(patentId: string): Promise<Patent> {
+        // compile the url for retrieving the patent
+        const url = process.env.PATENT_API_URL.concat(PatentsService.epoDocEndpoint).concat(patentId).concat('/biblio');
 
         const patentResponse = await this.sendOpsRequest<OpsPatent>(url, {}, 'get');
+
+        // return the processed version of the patent
         return PatentsServiceHelper.processPatent(
             patentResponse.data['ops:world-patent-data']['exchange-documents']['exchange-document'],
         );
@@ -40,13 +49,14 @@ export class PatentsService {
 
     /**
      * Queries the patent API with the provided search terms
+     *
      * @param terms The terms which will be used for querying
-     * @param page The page that should be queried
-     * @param languages
-     * @param country
-     * @param date
-     * @param inventor
-     * @param applicant
+     * @param page  The page that should be queried
+     * @param languages The preferred language(s)
+     * @param country   The country filter
+     * @param date      The date filter
+     * @param inventor  The inventor filter
+     * @param applicant The applicant filter
      */
     public async query(
         terms: string[],
@@ -57,6 +67,7 @@ export class PatentsService {
         inventor?: string,
         applicant?: string,
     ): Promise<QueryResult> {
+        // compile query string
         const queryString = PatentsServiceHelper.getQueryString(
             terms,
             PatentsService.patentEndpoint,
@@ -69,20 +80,25 @@ export class PatentsService {
         );
 
         const response = await this.sendOpsRequest<PatentQueryResponse>(queryString, {}, 'get');
+
+        // return processed patent
         return PatentsServiceHelper.processQuery(response.data, languages);
     }
 
     /**
      * Queries the patent family of a patent
-     * @param patentId
+     *
+     * @param patentId  The patentId for which to query the family
      */
     public async queryFamily(patentId: string) {
+        // compile the query string
         const queryString = process.env.PATENT_API_URL.concat(PatentsService.familyEndpoint)
             .concat(patentId)
             .concat('/biblio');
 
         const opsResponse = await this.sendOpsRequest<PatentFamilyResponse>(queryString, {}, 'get');
 
+        // convert family response to patent response to simplify processing
         const converted = {
             'ops:world-patent-data': {
                 'ops:biblio-search': {
@@ -96,30 +112,37 @@ export class PatentsService {
             },
         } as PatentQueryResponse;
 
+        // return processed patents
         return PatentsServiceHelper.processQuery(converted);
     }
 
     /**
-     * Returns available images for a patent
-     * @param patentId
+     * Returns document information for a patent
+     *
+     * @param patentId  The patent id
      */
     public async queryDocuments(patentId: string): Promise<DocumentInformation[]> {
-        const queryUrl = `${process.env.PATENT_API_URL}${PatentsService.imageEndpoint}${patentId}/images`;
+        // compile query string for the request
+        const queryUrl = `${process.env.PATENT_API_URL}${PatentsService.epoDocEndpoint}${patentId}/images`;
 
         const opsResponse = await this.sendOpsRequest<OpsImageQueryResponse>(queryUrl, {}, 'get');
 
+        // return processed response
         return PatentsServiceHelper.processImageQuery(opsResponse.data);
     }
 
     /**
-     * Gets a document from the OPS rest services
-     * @param url
-     * @param contentType
-     * @param range
+     * Retrieves a document from the OPS rest services
+     *
+     * @param url   The target-url of the document
+     * @param contentType   The content type which should be requested
+     * @param range The page which should be retrieved (only 1 page possible)
      */
     public async getDocument(url: string, contentType: string, range: number) {
+        // compile query url
         const queryUrl = `${process.env.PATENT_API_URL}/rest-services/${url}`;
 
+        // return the binary document response
         return this.sendOpsRequest(
             queryUrl,
             {
@@ -137,12 +160,13 @@ export class PatentsService {
 
     /**
      * Sends an ops request. It will also attempt to renew the access token once expired
-     * @param endpoint
-     * @param config
-     * @param requestType
-     * @param isSecondAttempt
-     * @param accept
-     * @param acceptEncoding
+     *
+     * @param endpoint  The endpoint for the request
+     * @param config    The config for the fetch request
+     * @param requestType   The request type (get, post, ...)
+     * @param isSecondAttempt   Optional parameter for secondary requests (if auth has to be renewed)
+     * @param accept    The accept header (defaults to 'application/json')
+     * @param acceptEncoding    The accept-encoding (defaults to 'gzip, deflate, br')
      * @private
      */
     private async sendOpsRequest<T>(
@@ -168,6 +192,8 @@ export class PatentsService {
             };
 
             const response = await lastValueFrom(this.httpService[requestType]<T>(endpoint, config));
+
+            // return response data and headers
             return { data: response.data as T, headers: response.headers };
         } catch (error) {
             // if access token is expired we will attempt to renew it and send the request again
@@ -178,8 +204,12 @@ export class PatentsService {
                 }
 
                 console.log('Authentication error. Attempting to refresh the token');
+
+                // renew access token
                 this.authData = await this.getAccessToken();
-                return this.sendOpsRequest<T>(endpoint, config, requestType, true);
+
+                // try request again
+                return this.sendOpsRequest<T>(endpoint, config, requestType, true, accept, acceptEncoding);
             }
             throw error;
         }
@@ -204,6 +234,7 @@ export class PatentsService {
         };
 
         try {
+            // sending auth-request
             const response = await lastValueFrom(
                 this.httpService.post<AuthResponse>(
                     `${process.env.PATENT_API_URL}/auth/accesstoken`,
@@ -213,6 +244,8 @@ export class PatentsService {
                     },
                 ),
             );
+
+            // returning the auth-response
             return response.data as AuthResponse;
         } catch (error) {
             console.error(`Error while authenticating with ${process.env.PATENT_API_URL}`);
